pyproject.toml
[tool.poetry]
authors = [ "SeparateRecords <me@rob.ac>" ]
description = "Discord bot for keeping people in sync by counting you in."
license = "ISC"
name = "count"
readme = "README.md"
version = "2.4.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = { version = "^1.5.0", extras = [ "voice" ] }
click = "^7.1.2"
loguru = "^0.5.3"
pydub = "^0.24.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
vermin = "^0.10.5"
"discord.py-stubs" = "^1.5.0"

[tool.poetry.scripts]
count-bot = "count.__main__:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
